Cleaning up AI slop: - override missing dbconnection members - remove redundant methods - configure statement capture - update docs - adjust file structure
