+++ src/kasthack.telemetry.dbconnection.di/TelemetryDbConnectionServiceCollectionExtensions.cs
@@ -1,7 +1,9 @@
-using kasthack.telemetry.dbconnection;
+using kasthack.telemetry.dbconnection.Options;
+
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+
 using System.Data.Common;
 
 namespace kasthack.telemetry.dbconnection.di;
