+++ src/kasthack.telemetry.dbconnection.slnx
@@ -1,4 +1,9 @@
 <Solution>
+  <Folder Name="/Solution Items/">
+    <File Path=".editorconfig" />
+    <File Path="Directory.Build.props" />
+    <File Path="Directory.Packages.props" />
+  </Folder>
   <Project Path="kasthack.telemetry.dbconnection.di/kasthack.telemetry.dbconnection.di.csproj" />
   <Project Path="kasthack.telemetry.dbconnection.ef/kasthack.telemetry.dbconnection.ef.csproj" />
   <Project Path="kasthack.telemetry.dbconnection/kasthack.telemetry.dbconnection.csproj" />
