+++ src/kasthack.telemetry.dbconnection/Constants/DbSemanticConventions.cs
@@ -1,4 +1,4 @@
-namespace kasthack.telemetry.dbconnection;
+namespace kasthack.telemetry.dbconnection.Constants;
 
 /// <summary>OpenTelemetry semantic convention tag names for database operations.</summary>
 internal static class DbSemanticConventions
