+++ src/kasthack.telemetry.dbconnection/Options/ConnectionManagementTracking.cs
@@ -1,4 +1,4 @@
-namespace kasthack.telemetry.dbconnection;
+namespace kasthack.telemetry.dbconnection.Options;
 
 /// <summary>Controls which connection management operations are instrumented with traces and metrics.</summary>
 [Flags]
@@ -9,7 +9,7 @@ public enum ConnectionManagementTracking
     /// <summary>Instrument open operations.</summary>
     Open = 1,
     /// <summary>Instrument close operations.</summary>
-    Close = 2,
+    Close = 1 << 1,
     /// <summary>Instrument all connection management operations.</summary>
     All = Open | Close,
 }
